+++ modules/crypto/pbkdf2v2.c
@@ -10,7 +10,7 @@
  */
 
 #include <atheme.h>
-#include <atheme/pbkdf2v2.h>
+#include <atheme/pbkdf2.h>
 
 #ifdef HAVE_LIBIDN
 #  include <stringprep.h>